presence_check/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.8
django-filter==21.1
django-qr-code==2.2.0
djangorestframework==3.12.4
image==1.5.33
Markdown==3.3.4
Pillow==8.3.2
pycodestyle==2.8.0
pytz==2021.3
qrcode==7.3.1
segno==1.3.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
